Don't show highlight when confirming :substitute
Using :s///c highlights searches to confirm them. Showing our highlight
makes it easy to confuse the two.

This is especially confusing when using IncSearch as the Searchlight
highlight because there are two identically highlighted matches.

'r' for mode() is 'Hit-enter prompt', which essentially means vim is
asking for input outside of the buffer (when it wouldn't be appropriate
to highlight).